+++ src/DependencyInjection/SentryExtension.php
@@ -27,9 +27,11 @@ use Sentry\SentryBundle\Tracing\Doctrine\DBAL\TracingDriverMiddleware;
 use Sentry\SentryBundle\Tracing\Twig\TwigTracingExtension;
 use Sentry\Serializer\RepresentationSerializer;
 use Sentry\Serializer\Serializer;
+use Sentry\Transport\TransportFactoryInterface;
 use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Component\Cache\CacheItem;
 use Symfony\Component\Config\FileLocator;
+use Symfony\Component\Debug\Exception\FatalErrorException;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Loader;
@@ -42,7 +44,7 @@ final class SentryExtension extends ConfigurableExtension
     /**
      * {@inheritdoc}
      */
-    public function getXsdValidationBasePath()
+    public function getXsdValidationBasePath(): string
     {
         return __DIR__ . '/../Resources/config/schema';
     }
@@ -50,7 +52,7 @@ final class SentryExtension extends ConfigurableExtension
     /**
      * {@inheritdoc}
      */
-    public function getNamespace()
+    public function getNamespace(): string
     {
         return 'https://sentry.io/schema/dic/sentry-symfony';
     }
@@ -124,6 +126,13 @@ final class SentryExtension extends ConfigurableExtension
             ->setPublic(false)
             ->setArgument(0, new Reference('sentry.client.options'));
 
+        $loggerReference = null === $config['logger']
+            ? new Reference(NullLogger::class, ContainerBuilder::IGNORE_ON_INVALID_REFERENCE)
+            : new Reference($config['logger']);
+
+        $factoryBuilderDefinition = $container->getDefinition(TransportFactoryInterface::class);
+        $factoryBuilderDefinition->setArgument('$logger', $loggerReference);
+
         $clientBuilderDefinition = (new Definition(ClientBuilder::class))
             ->setArgument(0, new Reference('sentry.client.options'))
             ->addMethodCall('setSdkIdentifier', [SentryBundle::SDK_IDENTIFIER])
@@ -131,7 +140,7 @@ final class SentryExtension extends ConfigurableExtension
             ->addMethodCall('setTransportFactory', [new Reference($config['transport_factory'])])
             ->addMethodCall('setSerializer', [$serializer])
             ->addMethodCall('setRepresentationSerializer', [$representationSerializerDefinition])
-            ->addMethodCall('setLogger', [null !== $config['logger'] ? new Reference($config['logger']) : new Reference(NullLogger::class, ContainerBuilder::IGNORE_ON_INVALID_REFERENCE)]);
+            ->addMethodCall('setLogger', [$loggerReference]);
 
         $container
             ->setDefinition('sentry.client', new Definition(Client::class))
@@ -265,7 +274,14 @@ final class SentryExtension extends ConfigurableExtension
             // Prepend this integration to the beginning of the array so that
             // we can save some performance by skipping the rest of the integrations
             // if the error must be ignored
-            array_unshift($integrations, new Definition(IgnoreErrorsIntegration::class, [['ignore_exceptions' => [FatalError::class]]]));
+            array_unshift($integrations, new Definition(IgnoreErrorsIntegration::class, [
+                [
+                    'ignore_exceptions' => [
+                        FatalError::class,
+                        FatalErrorException::class,
+                    ],
+                ],
+            ]));
         }
 
         return $integrations;