+++ src/Tracing/Doctrine/DBAL/TracingStatementForV3.php
@@ -1,0 +1,45 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sentry\SentryBundle\Tracing\Doctrine\DBAL;
+
+use Doctrine\DBAL\Driver\Result;
+use Doctrine\DBAL\Driver\Statement;
+use Doctrine\DBAL\ParameterType;
+use Sentry\Tracing\SpanContext;
+
+/**
+ * @internal
+ */
+final class TracingStatementForV3 extends AbstractTracingStatement implements Statement
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function bindValue($param, $value, $type = ParameterType::STRING): bool
+    {
+        return $this->decoratedStatement->bindValue($param, $value, $type);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function bindParam($param, &$variable, $type = ParameterType::STRING, $length = null): bool
+    {
+        return $this->decoratedStatement->bindParam($param, $variable, $type, $length);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function execute($params = null): Result
+    {
+        $spanContext = new SpanContext();
+        $spanContext->setOp(self::SPAN_OP_STMT_EXECUTE);
+        $spanContext->setDescription($this->sqlQuery);
+        $spanContext->setTags($this->spanTags);
+
+        return $this->traceFunction($spanContext, [$this->decoratedStatement, 'execute'], $params);
+    }
+}