+++ src/aliases.php
@@ -4,16 +4,19 @@ declare(strict_types=1);
 
 namespace Sentry\SentryBundle;
 
-use Doctrine\DBAL\Driver\ExceptionConverterDriver as LegacyExceptionConverterDriverInterface;
 use Doctrine\DBAL\Driver\Middleware as DoctrineMiddlewareInterface;
+use Doctrine\DBAL\Result;
 use Sentry\SentryBundle\EventListener\ErrorListenerExceptionEvent;
 use Sentry\SentryBundle\EventListener\RequestListenerControllerEvent;
 use Sentry\SentryBundle\EventListener\RequestListenerRequestEvent;
 use Sentry\SentryBundle\EventListener\RequestListenerResponseEvent;
 use Sentry\SentryBundle\EventListener\RequestListenerTerminateEvent;
 use Sentry\SentryBundle\EventListener\SubRequestListenerRequestEvent;
-use Sentry\SentryBundle\Tracing\Doctrine\DBAL\Compatibility\ExceptionConverterDriverInterface;
 use Sentry\SentryBundle\Tracing\Doctrine\DBAL\Compatibility\MiddlewareInterface;
+use Sentry\SentryBundle\Tracing\Doctrine\DBAL\TracingDriverForV2;
+use Sentry\SentryBundle\Tracing\Doctrine\DBAL\TracingDriverForV3;
+use Sentry\SentryBundle\Tracing\Doctrine\DBAL\TracingStatementForV2;
+use Sentry\SentryBundle\Tracing\Doctrine\DBAL\TracingStatementForV3;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
 use Symfony\Component\HttpKernel\Event\ExceptionEvent;
 use Symfony\Component\HttpKernel\Event\FilterControllerEvent;
@@ -28,72 +31,64 @@ use Symfony\Component\HttpKernel\Kernel;
 
 if (version_compare(Kernel::VERSION, '4.3.0', '>=')) {
     if (!class_exists(ErrorListenerExceptionEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(ExceptionEvent::class, ErrorListenerExceptionEvent::class);
     }
 
     if (!class_exists(RequestListenerRequestEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(RequestEvent::class, RequestListenerRequestEvent::class);
     }
 
     if (!class_exists(RequestListenerControllerEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(ControllerEvent::class, RequestListenerControllerEvent::class);
     }
 
     if (!class_exists(RequestListenerResponseEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(ResponseEvent::class, RequestListenerResponseEvent::class);
     }
 
     if (!class_exists(RequestListenerTerminateEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(TerminateEvent::class, RequestListenerTerminateEvent::class);
     }
 
     if (!class_exists(SubRequestListenerRequestEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(RequestEvent::class, SubRequestListenerRequestEvent::class);
     }
 } else {
     if (!class_exists(ErrorListenerExceptionEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(GetResponseForExceptionEvent::class, ErrorListenerExceptionEvent::class);
     }
 
     if (!class_exists(RequestListenerRequestEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(GetResponseEvent::class, RequestListenerRequestEvent::class);
     }
 
     if (!class_exists(RequestListenerControllerEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(FilterControllerEvent::class, RequestListenerControllerEvent::class);
     }
 
     if (!class_exists(RequestListenerResponseEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(FilterResponseEvent::class, RequestListenerResponseEvent::class);
     }
 
     if (!class_exists(RequestListenerTerminateEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(PostResponseEvent::class, RequestListenerTerminateEvent::class);
     }
 
     if (!class_exists(SubRequestListenerRequestEvent::class, false)) {
-        /** @psalm-suppress UndefinedClass */
         class_alias(GetResponseEvent::class, SubRequestListenerRequestEvent::class);
     }
 }
 
 if (!interface_exists(DoctrineMiddlewareInterface::class)) {
-    /** @psalm-suppress UndefinedClass */
     class_alias(MiddlewareInterface::class, DoctrineMiddlewareInterface::class);
 }
 
-if (!interface_exists(LegacyExceptionConverterDriverInterface::class)) {
-    /** @psalm-suppress UndefinedClass */
-    class_alias(ExceptionConverterDriverInterface::class, LegacyExceptionConverterDriverInterface::class);
+if (!class_exists('Sentry\SentryBundle\Tracing\Doctrine\DBAL\TracingStatement')) {
+    if (class_exists(Result::class)) {
+        class_alias(TracingStatementForV3::class, 'Sentry\\SentryBundle\\Tracing\\Doctrine\\DBAL\\TracingStatement');
+        class_alias(TracingDriverForV3::class, 'Sentry\\SentryBundle\\Tracing\\Doctrine\\DBAL\\TracingDriver');
+    } elseif (interface_exists(Result::class)) {
+        class_alias(TracingStatementForV2::class, 'Sentry\\SentryBundle\\Tracing\\Doctrine\\DBAL\\TracingStatement');
+        class_alias(TracingDriverForV2::class, 'Sentry\\SentryBundle\\Tracing\\Doctrine\\DBAL\\TracingDriver');
+    }
 }