+++ tests/Tracing/Doctrine/DBAL/ServerInfoAwareConnectionStub.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sentry\SentryBundle\Tests\Tracing\Doctrine\DBAL;
+
+use Doctrine\DBAL\Driver\Connection;
+use Doctrine\DBAL\Driver\ServerInfoAwareConnection;
+
+interface ServerInfoAwareConnectionStub extends Connection, ServerInfoAwareConnection
+{
+}